refactor: split Game.ts (2135 lines) into focused modules
Extracts mapper functions, autocomplete cache, profile mapping, and
IGDB sync logic into four new helper modules. Game.ts retains its full
public static API via thin delegates; no external import sites changed.

- GameMappers.ts: row/API mapper functions and API data types
- GameAutocompleteCache.ts: autocomplete cache state and helpers
- GameProfileMapper.ts: profile API types and mapGameProfileFromApi
- GameIgdbSync.ts: IGDB sync cluster (importGameFromIgdb, saveFullGameMetadata, etc.)

Dead code removed: buildReleaseSignature (was never called).

Closes #909
